added first_name to fields
I added the first_name variable to the fields of the view :me of the UserBlueprint so that I could access that specific variable name from the frontend.  And while I had access to the data within the table, I did not have access to SHOW it in the frontend until I added it in.

In other words, the reason why is because I didn't sent the entire object to the frontend - I only sent specific fields to the frontend.