added in params
We had to add in :med_id, :pet_id, and :vet_id because we could not create a prescription without these specific references to the respective tables in our database.

Since the prescription required all of the model params AND the references, we had to include them as params.

When i created the prescriptions table, my requirements listed were that med_id, vet_id, pet_id MUST be specified (not nullable).  So these parameters were absolutely required to create a prescription.